Change urls to /thing/<judgment>
where thing might be metadata, judgment, lock.

This is because it's not easy to determine whether /ewca/2022/4/lock is
a judgment URI which contains the word lock, or whether that's outside the
judgment's name; we give everything unique names.

Also we use {judgmentUri:path} because whilst the OpenAPI spec
doesn't allow multiple parts of a path to be used together, FastAPI does.