Remove default values for parameters which are paths, compatibility fix.
When upgrading to FastAPI 0.100 we got errors like this

```
src/openapi_server/apis/reading_api.py:61: in <module>
    judgmentUri: str = Path(None, description=""),
/opt/hostedtoolcache/Python/3.9.17/x64/lib/python3.9/site-packages/fastapi/param_functions.py:54: in Path
    return params.Path(
/opt/hostedtoolcache/Python/3.9.17/x64/lib/python3.9/site-packages/fastapi/params.py:178: in __init__
    assert default is ..., "Path parameters cannot have a default value"
E   AssertionError: Path parameters cannot have a default value
```

It appears it is no longer valid to have an optional value for a parameter to a function which is
provided by a path (e.g. `@router.patch("/judgment/{judgmentUri:path}"`)

This should always have a value, anyway.